requirements.txt
python-telegram-bot
vnstock
git+https://github.com/phamdinhkhanh/vnquant.git@a8272d36cb4368dc1cc85e3a79e270bfa8b09848
TA-Lib
httpx
